Fixes issue with wrong abstract function name
The trait-methods relied on a method that was not defined in the trait.
Instead a different method was defined so users had to implement an
unused method as well as a method that wasn't described by the contract.

That is now fixed.

Apart from that there's also tests now to prohibit such things in the
future.